Create script for changing a card's BSP version
This bash script installs, updates and flash a BSP version onto a
card. It also reboots the machine to ensure the new version is used.
The script requires an input of the BSP version found in
/mnt/cloud/binaries/nfp-bsp/interim/deb/
Thus, it is suitable for Ubuntu machines and reloads the driver
with the `insmod` command.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>