Create extensive script for changing BSP veresion
This bash script installs, updates and flash a BSP version onto a
specific card using its PCI address. It also reboots the machine to
ensure the new version is used.
The script requires the following inputs:
- last part of path to bsp version e.g. /interim/deb/
- PCI address e.g. 0000:07:00.0
- fw file name e.g. nic_AMDA0097-0001_2x40.nffw
- bsp version e.g. 22.10~00111.main.282e56a-0.bionic_amd64.deb
Thus, it is suitable for Ubuntu machines. It reloads the driver and
firmware using a soft reset and flashing the fw to the PCI address.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>